requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.9.1
streamlit==1.28.0